Reorganize files and add services and models directories
